FE-13: Enable deployment to and viewing as gh page site
In order to allow the site to be tested, it is adapted to be deployed
and viewed as a github pages site. For this, primarily the routing
and navigation needed to be updated, since github pages does not
support BrowserRouter, only HashRouter from react-router, and
automatically adds a base route of the project name to the deployed
pages.