FE-14: Correct styles to follow BEM convention and for consistency
Impose BEM convention on all CSS modules and adjust react files
accordingly. Change the styles variables and medias to be imported
via use statements and not import (deprecated). The media mixins
were also not able to import correctly - to do this the way they are
referenced is changed and the variables usages arealso changed in
order to remain consistent with this.